feat: pass themeData from content to layoutEditor
Passes the themeData from content and applies it when in
layoutEditor mode.

This required moving getThemeData to the UniversalReceivers.

This change makes the themes in layout editor update to match
the published theme instantly.
